fix risk classification
