+++ src/javascript/binary/common_functions/risk_classification.js
@@ -1,42 +1,5 @@
-var page = require('../base/page').page;
-
 var RiskClassification = (function() {
-  var financial_assessment_url = page.url.url_for('user/settings/assessmentws');
-  var renderRiskClassificationPopUp = function () {
-      if (window.location.pathname === '/user/settings/assessmentws') {
-        window.location.href = page.url.url_for('user/settingsws');
-        return;
-      }
-      $.ajax({
-          url: financial_assessment_url,
-          dataType: 'html',
-          method: 'GET',
-          success: function(riskClassificationText) {
-              if (riskClassificationText.includes('assessment_form')) {
-                  var payload = $(riskClassificationText);
-                  showRiskClassificationPopUp(payload.find('#assessment_form'));
-                  FinancialAssessmentws.LocalizeText();
-                  $('#risk_classification #assessment_form').removeClass('invisible')
-                                      .attr('style', 'text-align: left;');
-                  $('#risk_classification #high_risk_classification').removeClass('invisible');
-                  $('#risk_classification #heading_risk').removeClass('invisible');
-                  $("#risk_classification #assessment_form").on("submit",function(event) {
-                      event.preventDefault();
-                      FinancialAssessmentws.submitForm();
-                      return false;
-                  });
-              }
-          },
-          error: function(xhr) {
-              return;
-          }
-      });
-      $("#risk_classification #assessment_form").on("submit",function(event) {
-          event.preventDefault();
-          FinancialAssessmentws.submitForm();
-          return false;
-      });
-  };
+  'use strict';
 
   var showRiskClassificationPopUp = function (content) {
     if ($('#risk_classification').length > 0) {
@@ -61,7 +24,6 @@ var RiskClassification = (function() {
   };
 
   return {
-    renderRiskClassificationPopUp: renderRiskClassificationPopUp,
     cleanup: cleanup
   };
 }());
